Renamed to api and receive express app as dependency
