+++ services/centralized-ipfs-data-vault/test/api.get.test.ts
@@ -1,22 +1,14 @@
-import { Express } from 'express'
-import { ipfsPinnerProviderFactory, IpfsPinnerProvider, Entities } from '@rsksmart/ipfs-pinner-provider'
-import { setupDataVault } from '../src/data-vault'
+import express, { Express } from 'express'
+import { ipfsPinnerProviderFactory, IpfsPinnerProvider } from '@rsksmart/ipfs-pinner-provider'
+import { setupDataVault } from '../src/api'
 import request from 'supertest'
-import { deleteDatabase } from '../../../packages/ipfs-cpinner-provider/test/util'
-import { Connection, createConnection } from 'typeorm'
+import { Connection } from 'typeorm'
+import { createSqliteConnection, deleteDatabase } from './util'
 
 const ipfsEndpoint = 'http://localhost:5001'
 
-const createSqliteConnection = (database: string) => createConnection({
-  type: 'sqlite',
-  database,
-  entities: Entities,
-  logging: false,
-  dropSchema: true,
-  synchronize: true
-})
-
-async function testContentIsAccessible(
+async function testContentIsAccessible (
+  app: Express,
   database: string,
   { did, key }: { did: string, key: string },
   arrange: (ipfsPinnerProvider: IpfsPinnerProvider) => Promise<any>,
@@ -25,7 +17,7 @@ async function testContentIsAccessible(
   // setup
   const dbConnection = await createSqliteConnection(database)
   const ipfsPinnerProvider = await ipfsPinnerProviderFactory(dbConnection, ipfsEndpoint)
-  const app = setupDataVault(ipfsPinnerProvider)
+  setupDataVault(app, ipfsPinnerProvider)
 
   // arrange
   await arrange(ipfsPinnerProvider)
@@ -40,9 +32,11 @@ async function testContentIsAccessible(
 }
 
 const testSingleContentIsAccessible = (
+  app: Express,
   database: string,
   { did, key, content }: { did: string, key: string, content: string }
 ) => testContentIsAccessible(
+  app,
   database,
   { did, key },
   (ipfsPinnerProvider: IpfsPinnerProvider) => ipfsPinnerProvider.create(did, key, content),
@@ -55,14 +49,14 @@ describe('GET', function (this: {
   ipfsPinnerProvider: IpfsPinnerProvider,
   app: Express
 }) {
+  beforeEach(() => { this.app = express() })
   afterEach(() => deleteDatabase(this.dbConnection, this.database))
 
   describe('tdd', () => {
-
     test('get content from given did and key', async () => {
       this.database = 'test-1.service.get.sqlite'
 
-      this.dbConnection = await testSingleContentIsAccessible(this.database, {
+      this.dbConnection = await testSingleContentIsAccessible(this.app, this.database, {
         did: 'did:ethr:rsk:testnet:0xce83da2a364f37e44ec1a17f7f453a5e24395c70',
         key: 'ExistingKey',
         content: 'This is the content'
@@ -72,7 +66,7 @@ describe('GET', function (this: {
     test('get a different content from given did and key', async () => {
       this.database = 'test-2.service.get.sqlite'
 
-      this.dbConnection = await testSingleContentIsAccessible(this.database, {
+      this.dbConnection = await testSingleContentIsAccessible(this.app, this.database, {
         did: 'did:ethr:rsk:testnet:0xce83da2a364f37e44ec1a17f7f453a5e24395c70',
         key: 'ExistingKey',
         content: 'This is another content'
@@ -82,7 +76,7 @@ describe('GET', function (this: {
     test('get content from a different did', async () => {
       this.database = 'test-3.service.get.sqlite'
 
-      this.dbConnection = await testSingleContentIsAccessible(this.database, {
+      this.dbConnection = await testSingleContentIsAccessible(this.app, this.database, {
         did: 'did:ethr:rsk:testnet:0xf3d8a97f31d81ac42073e3c085c6dadd83cd1a79',
         key: 'ExistingKey',
         content: 'This is another content for another did'
@@ -92,7 +86,7 @@ describe('GET', function (this: {
     test('get content from a different key', async () => {
       this.database = 'test-4.service.get.sqlite'
 
-      this.dbConnection = await testSingleContentIsAccessible(this.database, {
+      this.dbConnection = await testSingleContentIsAccessible(this.app, this.database, {
         did: 'did:ethr:rsk:testnet:0xf3d8a97f31d81ac42073e3c085c6dadd83cd1a79',
         key: 'AnotherKey',
         content: 'This is a content for another key'
@@ -108,6 +102,7 @@ describe('GET', function (this: {
       const key = 'AnotherKey'
 
       this.dbConnection = await testContentIsAccessible(
+        this.app,
         this.database,
         { did, key },
         async () => { }, // no op
@@ -124,11 +119,13 @@ describe('GET', function (this: {
       const content2 = 'This is another content'
 
       this.dbConnection = await testContentIsAccessible(
+        this.app,
         this.database,
         { did, key },
         async (ipfsPinnerProvider: IpfsPinnerProvider) => {
           await ipfsPinnerProvider.create(did, key, content1)
-          await ipfsPinnerProvider.create(did, key, content2)}, // no op
+          await ipfsPinnerProvider.create(did, key, content2)
+        },
         [content1, content2]
       )
     })
