+++ services/centralized-ipfs-data-vault/test/util.ts
@@ -1,0 +1,23 @@
+import { Connection, createConnection } from 'typeorm'
+import fs from 'fs'
+import { Entities } from '@rsksmart/ipfs-pinner-provider'
+
+export const createSqliteConnection = (database: string) => createConnection({
+  type: 'sqlite',
+  database,
+  entities: Entities,
+  logging: false,
+  dropSchema: true,
+  synchronize: true
+})
+
+export const resetDatabase = async (dbConnection: Promise<Connection>) => {
+  await (await dbConnection).dropDatabase()
+  await (await dbConnection).synchronize()
+}
+
+export const deleteDatabase = (connection: Connection, database: string) => connection.close().then(() => {
+  if (fs.existsSync(database)) fs.unlinkSync(database)
+})
+
+export const getRandomString = (): string => Math.random().toString(36).substring(3, 11)
